Epilogue修改，examples文件修改

+++ examples/03_gemm/04_int8_cm_epi_gemm/scripts/verify_data.py
@@ -1,7 +1,7 @@
 import numpy as np
 import argparse
 import sys
-NP_DATA_TYPE = np.int32
+NP_DATA_TYPE = np.int8
 
 def compareOutputData(M, N, K):
     print("--- Test Begin ---\n")
